% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.glm.fit.R
\name{block.glm.fit}
\alias{block.glm.fit}
\title{Function to fit a generalized linear model with a
  particular block structure for the design matrix.}
\usage{
block.glm.fit(x, z, y, weights = rep(1, nobs), start = NULL,
  etastart = NULL, mustart = NULL, offset = rep(0, nobs),
  families = list(gaussian()), row.families = rep(1, nobs),
  control = list())
}
\arguments{
\item{x}{The X matrix lying on the diagonal blocks of the design.}

\item{z}{The Z matrix lying on the blocks in the right column of
the design.}

\item{y}{Responses}

\item{weights}{Case weights}

\item{start}{Starting coefficients, which must be given}

\item{etastart}{Starting linear predictors (not implemented)}

\item{mustart}{Starting fits for the mean of y}

\item{offset}{Offsets for the generalized linear model}

\item{families}{List of all the families that occur for any
observation.}

\item{row.families}{List of integers whose length is
\code{length(y)} between 1 and \code{length(families)}.  If
\code{row.families[i]} is 3, then row \code{i} uses the third
family in the list of families.}

\item{control}{Similar to \code{control} for \code{glm.fit}}
}
\description{
This function fits a GLM where each row has a different family
  and the block structure is like that described in the
  Proportional Observer Bias paper.
}
\author{
William Fithian
}

