% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.projection.R
\name{block.projection}
\alias{block.projection}
\title{Weighted least-squares solve with a certain block structure}
\usage{
block.projection(x, z, w, y, inverse.hessian = FALSE, wt.tol = 1e-30)
}
\arguments{
\item{x}{The X matrix lying on the diagonal blocks of the design.}

\item{z}{The Z matrix lying on the blocks in the right column of the design.}

\item{w}{Vector of weights.}

\item{y}{Vector to be projected.}

\item{inverse.hessian}{Should be FALSE.}

\item{wt.tol}{Used in internal computations.  Should be small.}
}
\description{
The function is called by \code{block.glm.fit} at each iteration
to do a computationally efficient step of iterative reweighted
least squares.
}
\author{
William Fithian
}

