% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{Linear "family" for rows.}
\usage{
linear(tol = 1e-10)
}
\arguments{
\item{tol}{Numerical tolerance.  Should be small.}
}
\description{
Allows us to encode a GLM with an extra linear term as a GLM.
}
\author{
William Fithian
}

