% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{multispeciesPP-package}
\alias{multispeciesPP-package}
\title{Continuous-Time Correlated Random Walk Models for Animal Movement}
\description{
Provides facilities for fitting continuous time correlated random
walk models to animal tracking data, and simulating tracks from the
fitted model.
}
\author{
William Fithian
}
\references{
Fithian, W., Elith, J., Hastie, T., & Keith, D. A. (2015). Bias
  correction in species distribution models: pooling survey and collection
  data for multiple species. Methods in Ecology and Evolution, 6(4), 424-438.
}

