% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multispeciesPP.R
\name{multispeciesPP}
\alias{multispeciesPP}
\title{Fit multispecies proportional bias models.}
\usage{
multispeciesPP(sdm.formula, bias.formula, PA, PO, BG, species = names(PO),
  species.PA = species, species.PO = species, quadrat.size = 1,
  region.size = 1, start = NULL, inverse.hessian = FALSE,
  penalty.l2.sdm = 0.1, penalty.l2.bias = 0.1,
  penalty.l2.intercept = 1e-04, weights = rep(1, n.species * nrow(x)),
  control = list())
}
\arguments{
\item{sdm.formula}{Formula for dependence of the species process
on environmental covariates}

\item{bias.formula}{Formula for dependence of the observer bias on
covariates}

\item{PA}{Presence-absence data, as a data frame}

\item{PO}{Presence-only data, as a list of data frames whose names
are the species}

\item{BG}{Background data, as a data frame}

\item{species}{Names of the species}

\item{species.PA}{Species contributing PA data}

\item{species.PO}{Species contributing PO data}

\item{quadrat.size}{Size of presence-absence quadrats}

\item{region.size}{Size of study region}

\item{start}{Optional argument giving initial coefficients.  These
should be in standardized form; i.e. they should correspond to a
model with the covariates centered and scaled.}

\item{inverse.hessian}{Not currently used}

\item{penalty.l2.sdm}{L2 penalty on coefficients for the species intensity}

\item{penalty.l2.bias}{L2 penalty on coefficients for the bias}

\item{penalty.l2.intercept}{L2 penalty on the intercepts}

\item{weights}{Observation weights}

\item{control}{Analogous to \code{control} for \code{glm.fit}}
}
\description{
Fit the model described in the Proportional Bias Model paper.
}
\examples{
n.pixel <- 1000
n.other.spec <- 20
spec.names <- letters[1:(n.other.spec+1)]

## Geographic covariates affecting species abundance
x <- matrix(rnorm(2*n.pixel),nrow=n.pixel)

## Geographic covariate causing selection bias (correlated with x1)
z <- scale(x[,1] + rnorm(n.pixel)*sqrt(.95^(-2)-1))

## Intercept and slopes for abundance rate
alpha <- c(-2,.3*rnorm(n.other.spec)-2)
beta <- cbind(c(1,-.5),matrix(rnorm(n.other.spec*2)/2,2))

## Intercept and slope for selection bias
gamma <- -4
delta <- -0.3

## PO data is impacted by selection bias
po.count <- matrix(rpois(n.pixel*(n.other.spec+1),
                         lambda=exp(rep(alpha,each=n.pixel)+x\%*\%beta+gamma+c(z)*delta)),
                   n.pixel,dimnames=list(NULL,spec.names))

PO.list <- lapply(spec.names,function(sp)
  data.frame(x1=x[,1],x2=x[,2],z=z)[rep(1:n.pixel,po.count[,sp]),])
names(PO.list) <- spec.names
BG <- data.frame(x1=x[1,],x2=x[,2],z=z)

## PA data is unbiased
n.sites <- 500
pa.samp <- sample(1:n.pixel,n.sites)
pa.count <- matrix(rpois(n.sites*(n.other.spec+1),
                         lambda=exp(rep(alpha,each=n.sites)+x[pa.samp,]\%*\%beta)),
                   n.sites,dimnames=list(NULL,spec.names))
table(pa.count>0)

## PA data doesn't need biasing covariates

PA <- cbind(data.frame(x1=x[pa.samp,1],x2=x[pa.samp,2]),as.data.frame(1*(pa.count>0)))

full.mod <- multispeciesPP(~x1+x2,~z,PA,PO.list,BG,region.size=n.pixel)

summary(full.mod)

plot(c(rbind(alpha,beta)),c(full.mod$species.coef[1:3,]),
     xlab="true coefficients",ylab="fitted coefficients")
abline(0,1)
}
\author{
William Fithian
}

