% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.multispeciesPP.R
\name{predict.multispeciesPP}
\alias{predict.multispeciesPP}
\title{Predict species intensities or biased species intensities
  based on a multispeciesPP model.}
\usage{
\method{predict}{multispeciesPP}(object, newdata, sdm = TRUE, bias = FALSE,
  species = colnames(object$fit.PA), dispersion = NULL, terms = NULL,
  na.action = na.pass, ...)
}
\arguments{
\item{object}{a fitted object of class \code{multispeciesPP}}

\item{newdata}{optionally, a data frame in which to look for
variables with which to predict. If omitted, the fitted linear
predictors are used.}

\item{sdm}{??}

\item{bias}{??}

\item{species}{vector of species names.}

\item{dispersion}{ignored.}

\item{terms}{ignored.}

\item{na.action}{ignored.}

\item{...}{ignored.}
}
\description{
Predict species intensities or biased species intensities based on
a multispeciesPP model.
}
\author{
William Fithian
}

