% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.multispeciesPP.R
\name{print.summary.multispeciesPP}
\alias{print.summary.multispeciesPP}
\title{Print summary table for multispeciesPP model}
\usage{
\method{print}{summary.multispeciesPP}(x, ...)
}
\arguments{
\item{x}{a summary of an multispeciesPP object.}

\item{...}{ignored.}
}
\description{
Print summary table for multispeciesPP model
}
\author{
William Fithian
}

