% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.multispeciesPP.R
\name{summary.multispeciesPP}
\alias{summary.multispeciesPP}
\title{Construct summary table for multispeciesPP model}
\usage{
\method{summary}{multispeciesPP}(object, standardized = TRUE, ...)
}
\arguments{
\item{object}{a fitted multispeciesPP object}

\item{standardized}{should the standardized or the raw
coefficients be reported.}

\item{...}{ignored.}
}
\description{
Construct summary table for multispeciesPP model
}
\author{
William Fithian
}

